functional layout parser
