+++ src/services/food-log/manager.ts
@@ -2,8 +2,8 @@ import { TFile, Vault, Notice, normalizePath } from 'obsidian';
 import { FoodItem } from '../../types/nutrition';
 import { PluginSettings } from '../../types/settings';
 import { FileUtils } from '../file-utils';
-import { LayoutGenerator } from '../layout-generator';
-import { ContentParser } from '../content-parser';
+import * as LayoutGenerator from '../layout-generator';
+import * as ContentParser from '../content-parser';
 
 export interface FoodLogDeps {
   vault: Vault;
@@ -64,10 +64,8 @@ export async function deleteFoodLogItem(
       throw new Error('Food log file not found for today');
     }
 
-    const contentParser = new ContentParser();
-
     await deps.vault.process(existingFile, (content) => {
-      const deleteResult = contentParser.deleteCardFromContent(content, itemToDelete);
+      const deleteResult = ContentParser.deleteCardFromContent(content, itemToDelete);
 
       if (!deleteResult.success) {
         throw new Error('Item not found in food log');
@@ -83,18 +81,17 @@ export async function deleteFoodLogItem(
 
 async function createNewFoodLog(deps: FoodLogDeps, path: string, foodItems: FoodItem[]): Promise<void> {
   const fileUtils = new FileUtils(deps.vault);
-  const layoutGenerator = new LayoutGenerator(deps.settings);
-  const contentParser = new ContentParser();
+  const goals = deps.settings.nutritionGoals;
 
   const today = fileUtils.getTodayString();
-  const totals = contentParser.calculateTotals(foodItems);
+  const totals = ContentParser.calculateTotals(foodItems);
 
   let content = `# 🍽️ Food Log ${today}\n\n`;
   content += `## 🥗 Today's Meals\n\n`;
-  content += layoutGenerator.generateCardLayout(foodItems);
+  content += LayoutGenerator.generateCardLayout(foodItems, goals);
   content += '\n<div class="food-placeholder"></div>\n\n';
-  content += layoutGenerator.generateCTAButtons('foodlog');
-  content += layoutGenerator.generateDailySummary(totals);
+  content += LayoutGenerator.generateCTAButtons('foodlog');
+  content += LayoutGenerator.generateDailySummary(totals, goals);
   content += '\n\n---\n\n';
   content += '*✨ Generated by AI Nutrition Tracker Plugin*\n';
 
@@ -102,8 +99,8 @@ async function createNewFoodLog(deps: FoodLogDeps, path: string, foodItems: Food
 }
 
 async function appendToExistingLog(deps: FoodLogDeps, file: TFile, foodItems: FoodItem[]): Promise<void> {
-  const layoutGenerator = new LayoutGenerator(deps.settings);
-  const newCard = layoutGenerator.generateCardLayout(foodItems);
+  const goals = deps.settings.nutritionGoals;
+  const newCard = LayoutGenerator.generateCardLayout(foodItems, goals);
   const placeholder = '<div class="food-placeholder"></div>';
 
   await deps.vault.process(file, (content) => {
@@ -133,13 +130,12 @@ async function replaceInExistingLog(
   newFoodItems: FoodItem[],
   originalEntry: { food: string; quantity: string; calories: number; protein: number; carbs: number; fat: number }
 ): Promise<void> {
-  const layoutGenerator = new LayoutGenerator(deps.settings);
-  const contentParser = new ContentParser();
-  const newCardContent = layoutGenerator.generateCardLayout(newFoodItems);
+  const goals = deps.settings.nutritionGoals;
+  const newCardContent = LayoutGenerator.generateCardLayout(newFoodItems, goals);
 
   try {
     await deps.vault.process(file, (content) => {
-      const replacement = contentParser.replaceCardInPosition(content, originalEntry, newCardContent);
+      const replacement = ContentParser.replaceCardInPosition(content, originalEntry, newCardContent);
 
       if (replacement.success) {
         return recalculateTotals(deps, replacement.content);
@@ -157,20 +153,19 @@ async function replaceInExistingLog(
 }
 
 function recalculateTotals(deps: FoodLogDeps, content: string): string {
-  const contentParser = new ContentParser();
-  const layoutGenerator = new LayoutGenerator(deps.settings);
+  const goals = deps.settings.nutritionGoals;
 
-  const foodItems = contentParser.extractFoodItemsFromContent(content);
-  const totals = contentParser.calculateTotals(foodItems);
+  const foodItems = ContentParser.extractFoodItemsFromContent(content);
+  const totals = ContentParser.calculateTotals(foodItems);
 
-  const newSummary = layoutGenerator.generateDailySummary(totals);
+  const newSummary = LayoutGenerator.generateDailySummary(totals, goals);
 
   // Find the summary card
   const summaryStart = content.indexOf('<div class="ntr-summary-card');
 
   if (summaryStart !== -1) {
     // Find the matching closing div for the summary card
-    const summaryBounds = contentParser.findCardBounds(content, summaryStart);
+    const summaryBounds = ContentParser.findCardBounds(content, summaryStart);
 
     if (summaryBounds.success) {
       // Replace existing summary
@@ -181,7 +176,7 @@ function recalculateTotals(deps: FoodLogDeps, content: string): string {
   }
 
   // No summary exists, add it (this shouldn't happen with placeholder approach)
-  const ctaButtons = layoutGenerator.generateCTAButtons('foodlog');
+  const ctaButtons = LayoutGenerator.generateCTAButtons('foodlog');
   return (
     content.trim() +
     '\n\n' +
